update schema via pre-push hook
This is an update from the synchronization script:
<https://github.com/hubverse-org/hubUtils/blob/main/.github/CONTRIBUTING.md#synchronizing-with-hubverse-orgschemas>

In https://github.com/hubverse-org/schemas/pull/122, the description for
the tasks schema was updated to clarify in the schema docs that multiple
target keys. This is a valid update since, while we use releases to
indicate a schema release, the hubs all pull the schema from the main
branch.

This also shows a bit of a weakness with git hooks: they only work if
people actually install them.

The reason I found this was because
https://github.com/hubverse-org/hubUtils/pull/218 modified the return
statements and the hook informed me that it was outdated. When I
updated it, the schemas updated.